feat: 删除长消息处理策略
