+++ config-template.py
@@ -224,26 +224,11 @@ show_prefix = False
 # 若您不需要此功能，请将force_delay_range设置为[0, 0]
 force_delay_range = [1.5, 3]
 
-# 应用长消息处理策略的阈值
-# 当回复消息长度超过此值时，将使用长消息处理策略
-blob_message_threshold = 256
-
-# 长消息处理策略
-# - "image": 将长消息转换为图片发送
-# - "forward": 将长消息转换为转发消息组件发送
-blob_message_strategy = "forward"
-
 # 允许等待
 # 同一会话内，是否等待上一条消息处理完成后再处理下一条消息
 # 若设置为False，若上一条未处理完时收到了新消息，将会丢弃新消息
 # 丢弃消息时的提示信息可以在tips.py中修改
 wait_last_done = True
-
-# 文字转图片时使用的字体文件路径
-# 当策略为"image"时生效
-#   若在Windows系统下，程序会自动使用Windows自带的微软雅黑字体
-#   若未填写或不存在且不是Windows，将禁用文字转图片功能，改为使用转发消息组件
-font_path = ""
 
 # 消息处理超时重试次数
 retry_times = 3
