+++ pkg/qqbot/message.py
@@ -6,7 +6,6 @@ import pkg.openai.session
 
 import pkg.plugin.host as plugin_host
 import pkg.plugin.models as plugin_models
-import pkg.qqbot.blob as blob
 import tips as tips_custom
