+++ pkg/qqbot/process.py
@@ -24,7 +24,6 @@ import pkg.qqbot.ratelimit as ratelimit
 
 import pkg.plugin.host as plugin_host
 import pkg.plugin.models as plugin_models
-import pkg.qqbot.blob as blob
 import tips as tips_custom
 
 processing = []
@@ -148,7 +147,6 @@ def process_message(launcher_type: str, launcher_id: int, text_message: str, mes
                                                  reply[0][:min(100, len(reply[0]))] + (
                                                      "..." if len(reply[0]) > 100 else "")))
                 reply = [reply[0]]
-                reply = blob.check_text(reply[0])
             else:
                 logging.info("回复[{}]消息".format(session_name))
