+++ pkg/qqbot/sources/nakuru.py
@@ -1,4 +1,6 @@
 import mirai
+from mirai.models.message import MessageComponent, MessageChain, Image, ForwardMessageNode
+from mirai.models.base import MiraiBaseModel
 
 from ..adapter import MessageSourceAdapter, MessageConverter, EventConverter
 import nakuru
@@ -10,7 +12,32 @@ import traceback
 import logging
 import json
 
-from pkg.qqbot.blob import Forward, ForwardMessageNode, ForwardMessageDiaplay
+from typing import List
+
+class ForwardMessageDiaplay(MiraiBaseModel):
+    title: str = "群聊的聊天记录"
+    brief: str = "[聊天记录]"
+    source: str = "聊天记录"
+    preview: List[str] = []
+    summary: str = "查看x条转发消息"
+    
+
+class Forward(MessageComponent):
+    """合并转发。"""
+    type: str = "Forward"
+    """消息组件类型。"""
+    display: ForwardMessageDiaplay
+    """显示信息"""
+    node_list: List[ForwardMessageNode]
+    """转发消息节点列表。"""
+    def __init__(self, *args, **kwargs):
+        if len(args) == 1:
+            self.node_list = args[0]
+            super().__init__(**kwargs)
+        super().__init__(*args, **kwargs)
+
+    def __str__(self):
+        return '[聊天记录]'
 
 
 class NakuruProjectMessageConverter(MessageConverter):
